python_utils/requirements.txt
pymongo
requests
configparser
pandas==1.0.3
numpy==1.18.2
elasticsearch==7.6.0
pypika==0.36.3
SQLAlchemy==1.3.11
pymssql==2.1.4
sqlalchemy-views==0.2.4
pyyaml